Move and redo SiegeTask and SiegePaymentTask
