+++ src/main/java/net/countercraft/movecraft/warfare/features/siege/tasks/SiegeTask.java
@@ -1,7 +1,8 @@
-package net.countercraft.movecraft.warfare.siege;
+package net.countercraft.movecraft.warfare.features.siege.tasks;
 
 import org.bukkit.scheduler.BukkitRunnable;
-import net.countercraft.movecraft.warfare.localisation.I18nSupport;
+
+import net.countercraft.movecraft.warfare.features.siege.Siege;
 
 public abstract class SiegeTask extends BukkitRunnable {
     protected final Siege siege;
